# build: refactor buildPageTree into a prebuild script