… lib/lib.js
… package.json