@matildepark 2022-05-05
All in all the deploy goes from 6m 17s to 5m 54s. Everything works, merging ...